requirements.txt
# EcoSort AI - Required Libraries

streamlit          # সুন্দর ওয়েব অ্যাপ তৈরি করার জন্য (app.py চালানোর জন্য)
tensorflow         # AI মডেল তৈরি ও ট্রেনিং করার জন্য (মূল লাইব্রেরি)
keras              # TensorFlow এর উপর সহজ লেয়ার (মডেল বানানো সহজ করে)
numpy              # ছবিকে নাম্বার অ্যারেতে কনভার্ট করার জন্য
pillow             # ছবি পড়া, resize করা এবং প্রসেস করার জন্য